账户

+++ 00-code/zl-maindir/zl-common-utils/src/main/java/com/ebeijia/zl/common/utils/tools/AmountUtil.java
@@ -52,8 +52,8 @@ public abstract class AmountUtil {
 	 * @param v2乘数
 	 * @return
 	 */
-	public static double mul(BigDecimal v1, BigDecimal v2) {
-		return v1.multiply(v2).doubleValue();
+	public static BigDecimal mul(BigDecimal v1, BigDecimal v2) {
+		return v1.multiply(v2);
 	}
 
 	/**
